requirements.txt
Django==1.5.1
Markdown==2.3.1
Pillow==2.0.0
South==0.7.6
amqp==1.0.10
anyjson==0.3.3
billiard==2.7.3.23
celery==3.0.17
django-bitcoin==0.1
django-coverage==1.2.2
django-crispy-forms==1.2.3
django-debug-toolbar==0.9.4
django-guardian==1.1.0.beta
django-pagination==1.0.7
django-taggit==0.10a1
django-thumbs==0.4
django-userena==1.2.0
easy-thumbnails==1.2
kombu==2.5.8
python-dateutil==1.5
requests==1.2.0
sorl-thumbnail==11.12
wsgiref==0.1.2
